+++ internal/rest/server_proposals.go
@@ -446,7 +446,7 @@ func convertProposalToInternal(pr *coreproposal.Proposal, di *internaldao.DAO) p
 		Body: []common.Content{
 			{
 				Type: common.Markdown,
-				Body: ipfs.ReplaceLinksInText(pr.Body),
+				Body: helpers.ReplaceInlineImages(ipfs.ReplaceLinksInText(pr.Body)),
 			},
 		},
 		Discussion:    pr.Discussion,